Fixed schematic not being placed in the correct location in some circumstances; Fixed forgotten light generation step
This fix places the map in the correct location and adds light pre-generation.

The light pre-gen step adds some time to the map placement, but makes loading the map during gameplay less resource intensive. We're doing it through emerge rather than when placing the schematics so that we don't crash the server through memory exhaustion. To disable this step, just comment out UBCMap:calculateLighting(startPos, endPos)